Module_3_Retrieval/YooXQuery/requirements.txt
pytorch
torchvision
cudatoolkit
pandas
numpy
matplotlib
tqdm
json

